+++ Native/OpenDDSWrapper/ParticipantService.h
@@ -41,4 +41,4 @@ EXTERN_METHOD_EXPORT
 long ParticipantService_GetScheduler();
 
 EXTERN_METHOD_EXPORT
-void ParticipantService_Scheduler(long scheduler);
+void ParticipantService_Scheduler(int scheduler);
